Completed Pizza Model, Controller, View and Route

+++ routes/web.php
@@ -11,17 +11,25 @@
 |
 */
 
+use App\User;
+
 Route::get('/', function () {
     return view('welcome');
 });
 
 // Pizza
-Route::get('/pizzas', 'PizzaController@index')->name('pizzas.index');
-Route::get('/pizzas/create', 'PizzaController@create')->name('pizzas.create');
+Route::get('/pizzas', 'PizzaController@index')->name('pizzas.index')->middleware('auth');
+Route::get('/order', 'PizzaController@create')->name('pizzas.create');
 Route::post('/pizzas', 'PizzaController@store')->name('pizzas.store');
-Route::get('/pizzas/{id}', 'PizzaController@show')->name('pizzas.show');
-Route::delete('/pizzas/{id}', 'PizzaController@destroy')->name('pizzas.destroy');
+Route::get('/pizzas/{id}', 'PizzaController@show')->name('pizzas.show')->middleware('auth');
+Route::delete('/pizzas/{id}', 'PizzaController@destroy')->name('pizzas.destroy')->middleware('auth');
 
-Auth::routes();
-
+// Auth
+if(count(User::all()) > 0){
+    Auth::routes([
+        'register' => false
+    ]);
+}else{
+    Auth::routes();
+}
 Route::get('/home', 'HomeController@index')->name('home');
